Move explore component to legacy-explore
