+++ frontend/scripts/pages/tool.page.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
-import ToolSelector from 'react-components/tool-selector';
 import Tool from 'react-components/tool';
 import { Provider } from 'react-redux';
 import BaseMarkup from 'html/base.ejs';
@@ -10,8 +9,6 @@ import CookieBanner from 'react-components/shared/cookie-banner';
 
 export const mount = (root, store) => {
   root.innerHTML = BaseMarkup();
-  const state = store.getState();
-  const { editing } = state.toolSelector;
   render(
     <Provider store={store}>
       <FiltersNav />
@@ -21,7 +18,7 @@ export const mount = (root, store) => {
 
   render(
     <Provider store={store}>
-      {editing && ENABLE_REDESIGN_PAGES ? <ToolSelector /> : <Tool />}
+      <Tool />
       <Feedback />
     </Provider>,
     document.getElementById('page-react-root')
