+++ frontend/scripts/react-components/legacy-explore/explore.reducer.js
