+++ frontend/scripts/react-components/legacy-explore/explore.scss
