+++ frontend/scripts/store.js
@@ -8,7 +8,7 @@ export {
 export { default as newsletter } from 'react-components/shared/newsletter/newsletter.reducer';
 export { default as staticContent } from 'react-components/static-content/static-content.reducer';
 export { default as team } from 'react-components/team/team.reducer';
-export { default as explore } from 'react-components/explore/explore.reducer';
+export { default as legacyExplore } from 'react-components/legacy-explore/explore.reducer';
 export { default as widgets } from 'react-components/widgets/widgets.reducer';
 export { default as dashboardRoot } from 'react-components/dashboard-root/dashboard-root.reducer';
 export {
